requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets>=13.0
pydantic>=2.0.0
python-multipart>=0.0.9

# Autenticación
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.0

# Ciencia de datos
numpy>=1.24.0
scipy>=1.10.0
h5py>=3.8.0
